+++ app/controllers/organizations_controller.rb
@@ -116,6 +116,9 @@ class OrganizationsController < HtmlController
     @pagy_users, @members = pagy apply_scopes(@organization.members, member_order_scope)
     @new_member = User.new
     @roles = Role::LOCAL_ROLES.keys
+
+    @selected_date = params[:selected_date] || Time.zone.today
+    @lesson_summaries = GroupLessonSummary.where(chapter_id: @organization.chapters, lesson_date: @selected_date)
   end
 
   def chapter_order_scope