+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 - Added ability to import students to an organization from `.csv` files
+- Added organization statistics in separate component when viewing an organization
 
 ## 0.37.0 - Student Enrollments Overhaul
 - Changed serializing and scoping for students to go through organization and by groups found in their enrollments